LayoutTests/inspector-protocol/page/deny-X-FrameOption-expected.txt

CONSOLE MESSAGE: Refused to display 'deny-X-FrameOption.html' in a frame because it set 'X-Frame-Options' to 'deny'.

deny-X-FrameOption.html:30:17
